backend/yacht_agents/requirements.txt
google-adk
anyio
python-dotenv
pydantic